AutoAnchor.java - damage calculations update - code cleanup
update MathUtils.java
- rayCast should now work
- code cleanup

update WorldUtils.java
- refactored some stuff
- code cleanup

update Watermark.java
- fixes

+++ src/main/java/org/snail/plus/utils/TPSSyncUtil.java
@@ -4,9 +4,9 @@ import static meteordevelopment.meteorclient.MeteorClient.mc;
 
 public class TPSSyncUtil {
 
-    public static double getSync(long time, java.util.concurrent.TimeUnit unit) {
+    public static double getSync(Double time, java.util.concurrent.TimeUnit unit) {
         int count;
-        long millis = unit.toMillis(time);
+        long millis = unit.toMillis(time.longValue());
         synchronized (TPSSyncUtil.class) {
             count = (int) (millis / mc.getServer().getAverageTickTime());
         }
